# backend-gymWeb/src/docs/swagger.yaml
basePath: null
definitions:
    AuthUser:
        description: ""
        properties:
            username:
                type: string
                description: ""
            password:
                type: string
                description: ""
        type: object
        required:
            - username
            - password
    IUser:
        description: ""
        properties:
            id:
                type: number
                format: double
                description: ""
            name:
                type: string
                description: ""
            lastname:
                type: string
                description: ""
            dni:
                type: number
                format: double
                description: ""
            dateofbirth:
                type: string
                description: ""
            phone:
                type: string
                description: ""
            email:
                type: string
                description: ""
            address:
                type: string
                description: ""
            city:
                type: string
                description: ""
            state:
                type: number
                format: double
                description: ""
        type: object
        required:
            - id
            - name
            - lastname
            - dni
            - dateofbirth
            - phone
            - email
            - address
            - city
            - state
    User:
        description: ""
        properties:
            id:
                type: number
                format: double
                description: ""
            name:
                type: string
                description: ""
            lastname:
                type: string
                description: ""
            dni:
                type: number
                format: double
                description: ""
            dateofbirth:
                type: string
                description: ""
            phone:
                type: string
                description: ""
            email:
                type: string
                description: ""
            address:
                type: string
                description: ""
            city:
                type: string
                description: ""
            state:
                type: number
                format: double
                description: ""
        type: object
        required:
            - id
            - name
            - lastname
            - dni
            - dateofbirth
            - phone
            - email
            - address
            - city
            - state
info:
    description: 'Paradigm API.'
    license:
        name: ISC
    title: paradigm-api
    version: 1.0.0
paths:
    /api/login:
        post:
            operationId: LoginPost
            produces:
                - text/html
            responses:
                '200':
                    description: Ok
                    schema:
                        type: string
            description: ""
            tags:
                - 'User login'
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: authUser
                    required: true
                    schema:
                        $ref: '#/definitions/AuthUser'
    /api/users:
        get:
            operationId: UserGetAll
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/IUser'
            description: ""
            tags:
                - User
            security:
                -
                    default:
                        - x-auth
            parameters: []
        post:
            operationId: UserNewUser
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/User'
            description: ""
            tags:
                - User
            security:
                -
                    default:
                        - x-auth
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: user
                    required: true
                    schema:
                        $ref: '#/definitions/User'
    '/api/users/{id}':
        get:
            operationId: UserGetById
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/IUser'
            description: ""
            tags:
                - User
            security:
                -
                    default:
                        - x-auth
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
        delete:
            operationId: UserDelete
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/User'
            description: ""
            tags:
                - User
            security:
                -
                    default:
                        - x-auth
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: number
                    format: double
    /api/users/edit:
        put:
            operationId: UserUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/User'
            description: ""
            tags:
                - User
            security:
                -
                    default:
                        - x-auth
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: user
                    required: true
                    schema:
                        $ref: '#/definitions/User'
swagger: '2.0'
securityDefinitions:
    default:
        type: apiKey
        name: x-auth
        in: header